requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
fastapi==0.115.7
filelock==3.17.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
llama-index-core==0.12.12
llama-index-embeddings-huggingface==0.5.1
llama-index-llms-ollama==0.5.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
ollama==0.4.7
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
